Move translation install from main.cpp to mymainwindow.cpp
This comment improves the language switch for fallback mode. From now,
the language switch option "Systemsprache" handles the fallback mode for
missing translations, too.
To support this feature, the initial translation installation is moved from
the main.cpp into the constructor of the mymainwindow.cpp to use the same
QTranslator "m_translator".