Fix override bug.
For me, the override file option does not work. The files are marked as
"NOT_SYNCHRONISED" but they will not downloaded. I figured out that the
line:
QFileInfo(directory, filename).lastModified().toMSecsSinceEpoch()/1000 <
currentElement->data(dateRole).toInt();
does not return true.

And so, I've replaced this line with the synchronisedRole check.
Now it works. Tested on linux.